Rename and reformat docs
