+++ docs/source/index.rst
@@ -4,10 +4,12 @@
 
 .. image:: images/logo.png
 
-``aguaclara`` is a Python package built by `AguaClara Cornell <http://aguaclara.cee.cornell.edu/>`_. It contains Python classes for designing AguaClara water treatment plants and functions for water treatment research.
+``aguaclara`` is a Python package built by `AguaClara Cornell
+<http://aguaclara.cee.cornell.edu/>`_. It contains Python classes for designing
+AguaClara water treatment plants and functions for water treatment research.
 
 .. toctree::
-    :maxdepth: 1
+    :maxdepth: 2
 
     core/core
     design/design
